feat: add downsampling query

+++ Client/Write/PointData.cs
@@ -163,25 +163,25 @@ namespace InfluxDB3.Client.Write
 
 
         /// <summary>
-        /// Gets the float field value associated with the specified name.
+        /// Gets the double field value associated with the specified name.
         /// If the field is not present, returns null.
         /// </summary>
         /// <param name="name">the field name</param>
-        /// <returns>The float field value or null</returns>
-        public float? GetFloatField(string name)
+        /// <returns>The double field value or null</returns>
+        public double? GetDoubleField(string name)
         {
-            return _values.GetFloatField(name);
+            return _values.GetDoubleField(name);
         }
 
         /// <summary>
-        /// Adds or replaces a float field.
+        /// Adds or replaces a double field.
         /// </summary>
         /// <param name="name">the field name</param>
         /// <param name="value">the field value</param>
         /// <returns>this</returns>
-        public PointData SetFloatField(string name, float value)
+        public PointData SetDoubleField(string name, double value)
         {
-            _values.SetFloatField(name, value);
+            _values.SetDoubleField(name, value);
             return this;
         }
